Add loop to jump special notes.
There are some pages with special notes that we need to skip to get the
petitions and what the celebrant says.

To avoid this, we skip the 'p' sections until we find one with two items
in the content. this is usually the Celebrant section (because of the
strong attribute for the word 'celebrant').

See
https://www.priestsforlife.org/liturgy/liturgicalresource-cycles.aspx?id=172
for an example.

Tested:
- Ran the script against the link above and one without special notes
(id=171) to confirm this works.